Prepare for codegen gRPC services
This makes GrpcMethod a dumb carrier type for an RPC endpoint. Two new methods on GrpcClient, newCall and newStreamingCall allow passing this model object to create a new call.
